Refactor tool base functionality
The working of each tool has been moved to each tool's own file instead of letting PaintWidget be a god class.
Tests were made only with transform tool as of now, and the way the tooling system was being implemented has been changed.
A base abstract class (effectively working as an interface) is going to command the selection tools. Pan tool doesn't come in this category.
OOP ftw.