Connecting tool-menus and tools (signals and slots)
This version crashes on attempting to draw bounds on a layer. This has to be checked/run through a debugger.
The connection has been made, and the tools now need to call the constructor of Abstract classes they are inheriting form instead of directly accessing tool to avoid the Q_OBJECT inclusion problem with virtual parents.