Creating new Layers using raw pointers
Reason being Layer* doesn't really need smart pointers, for the fact that they are being assigned to a parent immediately after they are created at a later point in their lifecycle as soon as they are pushed into the paintwidget (see PaintWidget::pushLayer(Layer*)). This however needs to be taken care of by explicitly or otherwise deleting the concerned parent item (as well as drawing). If that's a Q_Object, Qt will take care ofthe memory management for that object anyway, hence defeating the entire purpose of using a smart pointer here.